server/requirements.txt
fastapi
black
isort
websockets
pydantic
langchain
uvicorn
bs4
unstructured
libmagic
chromadb 
sentence_transformers 
InstructorEmbedding
python-dotenv
touch
pyllama 
llama-cpp-python
pyllamacpp
gpt4all
pygpt4all
tiktoken
accelerate
safetensors
torch
# xformers
bitsandbytes
einops
pypdf
